Rename testing.utils to testing.fixture.
